--- client/src/pages/Upload/index.tsx
@@ -1,2 +1,0 @@
-export { default } from "./Upload";
-export * from "./Upload";
