+++ server/handlers/file_handler.py
@@ -16,6 +16,7 @@ from dateparser.search import search_dates
 # supported course codes taken from
 # https://www.ucalgary.ca/pubs/calendar/current/course-desc-main.html
 
+
 def cache_course_codes():
     """
     Returns a list of all course codes
@@ -117,9 +118,7 @@ def extract_assessments(table):
             if not cell:  # skip empty cells
                 continue
             # first try dateparser
-            dates = search_dates(
-                cell, languages=["en"], settings={"REQUIRE_PARTS": ["day", "month"]}
-            )
+            dates = search_dates(cell, languages=["en"], settings={"REQUIRE_PARTS": ["day", "month"]})
             if dates:
                 source, date = dates[0]
             else:  # try datefinder if dateparser fails
@@ -133,9 +132,9 @@ def extract_assessments(table):
                 # Ignore dates that are too short to avoid false positives.
                 # The shortest a date can realistically be is 5 characters. e.g. Dec 1
                 continue
-            name = row[0].replace(
-                "\n", " "
-            )  # use the text in the first cell as the name
+            
+            # use the text in the first cell as the name if it exists
+            name = row[0].replace("\n", " ") if row[0] else "Unknown"
 
             weight = "unknown"
             for cell in row:
